+++ src/commonMain/kotlin/KSLog.kt
@@ -1,6 +1,7 @@
 package dev.inmo.kslog.common
 
 import dev.inmo.kslog.common.filter.filtered
+import dev.inmo.kslog.common.utils.plus
 
 
 enum class LogLevel {
@@ -110,3 +111,6 @@ fun KSLog(
 ).filtered { l, _, _ ->
     minLoggingLevel.ordinal <= l.ordinal
 }
+
+fun setDefaultKSLog(newDefault: KSLog) { KSLog.default = newDefault }
+fun addDefaultKSLog(newDefault: KSLog) { KSLog.default = KSLog.default + newDefault }